Don't use threads
Stop using threads, it's unnecessary, because DDG limits traffic, so
it's not useful.
Save bookmarks to the file as bangs are resolved.
Use the Location header to resolve a bang's URL.
Don't use the title from the bang's page.
If bangs have the same URL, add fragment identifier(s)(#) to the end of
URL of the bang.